% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpin_internals.R
\name{rpin.pin_internal}
\alias{rpin.pin_internal}
\title{Generate random  pins from birth distribution (given by logspline)}
\usage{
\method{rpin}{pin_internal}(x, distribution, ...)
}
\arguments{
\item{x}{length of the generated pin vector}

\item{distribution}{a logspline object with distribution to sample from}

\item{...}{arguments passed to birth2date}
}
\value{
pin vector of length x
}
\description{
Generate random  pins from birth distribution (given by logspline)
}

