% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpin_internals.R
\name{rpin_unique}
\alias{rpin_unique}
\title{Resample duplicated subset of pin vector to make all elements unique}
\usage{
rpin_unique(pin, recursive_fun, ...)
}
\arguments{
\item{pin}{a pin vector with possibly non unique elements}

\item{recursive_fun}{function to regenerate duplicated pins}

\item{...}{arguments pased to \code{recursive_fun}}
}
\description{
Resample duplicated subset of pin vector to make all elements unique
}
\keyword{internal}

